pyproject.toml
[tool.poetry]
name = "bumblebee"
version = "0.1.0"
description = ""
authors = ["rathourdevesh <rathourdevesh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "1.13.2"
annotated-types = "^0.7.0"
anyio = "^4.4.0"
bcrypt = "^4.2.0"
build = "^1.2.2"
cachecontrol = "^0.14.0"
cachetools = "^5.5.0"
certifi = "^2024.8.30"
cffi = "^1.17.1"
cfgv = "^3.4.0"
chardet = "^5.2.0"
charset-normalizer = "^3.3.2"
cleo = "^2.1.0"
click = "^8.1.7"
crashtest = "^0.4.1"
cssselect = "^1.2.0"
cssutils = "^2.11.1"
distlib = "^0.3.8"
dnspython = "^2.6.1"
dulwich = "^0.22.1"
email-validator = "^2.2.0"
exceptiongroup = "^1.2.2"
fastapi = "^0.114.1"
fastjsonschema = "^2.20.0"
filelock = "^3.16.0"
h11 = "^0.14.0"
httpcore = "^1.0.5"
httptools = "^0.6.1"
identify = "^2.6.0"
idna = "^3.8"
importlib-metadata = "^8.5.0"
iniconfig = "^2.0.0"
installer = "^0.7.0"
jaraco-classes = "^3.4.0"
keyring = "^25.3.0"
lxml = "^5.3.0"
mako = "^1.3.5"
markupsafe = "^2.1.5"
more-itertools = "^10.5.0"
msgpack = "^1.1.0"
mypy-extensions = "^1.0.0"
nodeenv = "^1.9.1"
packaging = "^24.1"
passlib = "^1.7.4"
pexpect = "^4.9.0"
pkginfo = "^1.11.1"
platformdirs = "^4.3.2"
pluggy = "^1.5.0"
premailer = "^3.10.0"
psycopg = "^3.2.1"
ptyprocess = "^0.7.0"
pycparser = "^2.22"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
pydantic-core = "^2.23.3"
pyjwt = "^2.9.0"
pyproject-hooks = "^1.1.0"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
rapidfuzz = "^3.9.7"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
shellingham = "^1.5.4"
six = "^1.16.0"
sniffio = "^1.3.1"
sqlalchemy = "^2.0.34"
starlette = "^0.38.5"
tomli = "^2.0.1"
tomlkit = "^0.13.2"
trove-classifiers = "^2024.9.12"
typing-extensions = "^4.12.2"
tzdata = "^2024.1"
urllib3 = "^2.2.3"
uvicorn = "^0.30.6"
uvloop = "^0.20.0"
virtualenv = "^20.26.4"
watchfiles = "^0.24.0"
websockets = "^13.0.1"
xattr = "^1.1.0"
zipp = "^3.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
